Cause gdal job fail from SSL SYSCALL error
GDAL sometimes has errors, but still returns exit code 0. We are
printint a warning log in this case.

However, we need to cause a failure if there is an SSL SYSCALL error in
the log, since that indicates a loss of connection of the database. That
could mean that a tile rasterization has been stopped early, so the tile
will be incorrect (since not all polygons will have been rasterized).