GTC-3025 Skip creation of tiles/extent.geojson if already there
This is to allow making progress on zoom_14 tile caches with a large
number of tiles, where the tiles/extent.geojson generation take a very
long time. This may happen with global 10m datasets.

Also, added back the missing cli() call that allows pixetl_prep.py to be
run locally, which can be used to run this long process without batch
timeouts.