Run fetching of meta-data of tile files in parallel
When generating the tiles.geojson file, the creation of the RasterSource
objects for each tile involves fetching meta-data from the tile files
from S3, and this can be quite slow when proceeding in serial. So, make
changes to fetch the meta-data in parallel using ProcessPoolExecutor.

For 16 workers, this reduces the fetch time on my local machine by
factor of 5-6, so reduce a nearly two-hour process for processing 2800
tiles to 20 minutes.